+++ Sources/Core/EmptyInitializable.swift
@@ -1,0 +1,6 @@
+/// Types conforming to this protocol can 
+/// be initialized with no arguments, allowing
+/// protocols to add static convenience methods.
+public protocol EmptyInitializable {
+    init()
+}
